feat: class Resources updated and moved
- `resources` and `package` were moved to `arcor2_execution_data`.
- Parameter `project_id` of the `Resources` class was removed.
