+++ src/python/arcor2/VERSION
@@ -1,1 +1,1 @@
-0.17.1
+0.18.0
