+++ src/python/arcor2_build/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.17.0] - WIP
+
+### Changed
+- Updated code generation for `Resources` (moved, now without parameters).
+
 ## [0.16.0] - 2021-06-11
 
 ### Changed
