+++ src/python/arcor2_build/VERSION
@@ -1,1 +1,1 @@
-0.16.0
+0.17.0
