+++ src/python/arcor2_execution/CHANGELOG.md
@@ -3,10 +3,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 
-## [0.15.0] - WIP
+## [0.16.0] - WIP
 
 ### Changed
-- Dependency on arcor2 0.17.1 (`constants` renamed to `parameters` in the `Project` model).
+- Dependency on `arcor2` 0.18.0.
+- Dependency on `arcor2_execution_data` 0.11.0.
+
+## [0.15.0] - 2021-06-14
+
+### Changed
+- Dependency on `arcor2` 0.17.0 (`constants` renamed to `parameters` in the `Project` model).
 
 ## [0.14.0] - 2021-05-21
