+++ src/python/arcor2_execution/VERSION
@@ -1,1 +1,1 @@
-0.15.0
+0.16.0
