+++ src/python/arcor2_execution/scripts/BUILD
@@ -2,5 +2,5 @@ python_library()
 
 arcor2_pex_binary(
     name = "execution",
-    dependencies = ["src/python/arcor2", "src/python/arcor2/object_types", "src/python/arcor2/parameter_plugins"],  # needed by the main script which run as subprocess of execution
+    dependencies = ["src/python/arcor2", "src/python/arcor2/object_types", "src/python/arcor2/parameter_plugins", "src/python/arcor2_execution_data"],  # needed by the main script which run as subprocess of execution
 )
