+++ src/python/arcor2_execution_data/VERSION
@@ -1,1 +1,1 @@
-0.10.0
+0.11.0
